move useSong()
