+++ projects/graphql-client/src/index.ts
@@ -6,3 +6,4 @@ export * from "./remote/queries/useArtists"
 export * from "./remote/queries/useGenres"
 export * from "./remote/queries/useTags"
 export * from "./remote/queries/useSongTypes"
+export * from "./remote/queries/useSong"
