+++ projects/shared-types/src/index.ts
@@ -4,3 +4,4 @@ export * from "./TimestampedResult"
 export * from "./Songs"
 export * from "./Users"
 export * from "./GeneratedTypes"
+export * from "./Fragements"
